// api/src/ai/tools/schema.ts
export const PrimaryKeyInputSchema = z.union([z.number(), z.string()]);
export const PrimaryKeyValidateSchema = z.union([z.number(), z.string()]);

export const ItemInputSchema = z.record(z.string(), z.any());
export const ItemValidateSchema = z.record(z.string(), z.any());

export const QueryInputSchema = z
    .object({
        fields: z.array(z.string()),
        sort: z.array(z.string()),
        filter: z.record(z.string(), z.any()),
        limit: z.number(),
        offset: z.number(),
        page: z.number(),
        search: z.string(),
        deep: z.record(z.string(), z.any()),
        alias: z.record(z.string(), z.string()),
        aggregate: z.object({
            count: z.array(z.string()).optional(),
            sum: z.array(z.string()).optional(),
            avg: z.array(z.string()).optional(),
            min: z.array(z.string()).optional(),
            max: z.array(z.string()).optional(),
        }),
        backlink: z.boolean(),
        version: z.string(),
        versionRaw: z.boolean(),
        export: z.string(),
        groupBy: z.array(z.string()),
    })
    .partial();

export const QueryValidateSchema = QueryInputSchema;
// ...
export const FileItemInputSchema = z
    .object({
        id: z.string(),
        storage: z.string(),
        filename_disk: z.string(),
        filename_download: z.string(),
        title: z.union([z.string(), z.null()]),
        type: z.union([z.string(), z.null()]),
        folder: z.union([z.string(), z.null()]),
        created_on: z.string(),
        uploaded_by: z.union([z.string(), z.null()]),
        uploaded_on: z.union([z.string(), z.null()]),
        modified_by: z.union([z.string(), z.null()]),
        modified_on: z.string(),
        charset: z.union([z.string(), z.null()]),
        filesize: z.number(),
        width: z.union([z.number(), z.null()]),
        height: z.union([z.number(), z.null()]),
        duration: z.union([z.number(), z.null()]),
        embed: z.union([z.string(), z.null()]),
        description: z.union([z.string(), z.null()]),
        location: z.union([z.string(), z.null()]),
        tags: z.union([z.string(), z.null()]),
        metadata: z.union([z.record(z.string(), z.any()), z.null()]),
        focal_point_x: z.union([z.number(), z.null()]),
        focal_point_y: z.union([z.number(), z.null()]),
        tus_id: z.union([z.string(), z.null()]),
        tus_data: z.union([z.record(z.string(), z.any()), z.null()]),
    })
    .partial();

export const FileItemValidateSchema = FileItemInputSchema;

export const FileImportItemInputSchema = z.object({
    url: z.string(),
    file: FileItemInputSchema,
});

export const FileImportItemValidateSchema = z.object({
    url: z.string(),
    file: FileItemValidateSchema,
});